sophia: add memory quota usage to statistics

+++ build/amalgamation.sh
@@ -1,28 +1,40 @@
 #!/bin/sh
 
+# sophia amalgamation build
+#
+
 process() {
-	output=$1
-	dir=$2
+	root=$1
+	output=$2
 	lib=$3
-	includes=`cat $dir/$lib | sed -n 's/#include <\(.*\)>/\1/p'`
+	dir=$4
+	includes=`cat $root/$dir/$lib | sed -n 's/#include <\(.*\)>/\1/p'`
 	for file in $includes; do
-		cat "$dir/$file" >> $output
+		cat "$root/$dir/$file" >> $output
 	done
-	files=`ls $dir/*.c`
+	files=`ls $root/$dir/*.c`
 	for file in $files; do
 		cat "$file" | grep -v "include" >> $output
 	done
 }
 
-output="sophia.c"
+if [ $# -ne 2 ]; then
+	echo "sophia amalgamation build."
+	echo "usage: $0 <sophia_root> <output>"
+	return 1
+fi
+
+root=$1
+output=$2
+
 rm -f $output
 touch $output
 
-process $output "rt" "libsr.h"
-process $output "version" "libsv.h"
-process $output "mvcc" "libsm.h"
-process $output "log" "libsl.h"
-process $output "database" "libsd.h"
-process $output "index" "libsi.h"
-process $output "repository" "libse.h"
-process $output "sophia" "libso.h"
+process $root $output "libsr.h" "rt"
+process $root $output "libsv.h" "version"
+process $root $output "libsm.h" "mvcc"
+process $root $output "libsl.h" "log"
+process $root $output "libsd.h" "database"
+process $root $output "libsi.h" "index"
+process $root $output "libse.h" "repository"
+process $root $output "libso.h" "sophia"
